+++ src/components/manage/Widgets/ArrayWidget.jsx
@@ -166,9 +166,10 @@ class ArrayWidget extends Component {
    */
   handleChange(selectedOption) {
     this.setState({ selectedOption });
+
     this.props.onChange(
       this.props.id,
-      selectedOption.map(item => item.value),
+      selectedOption ? selectedOption.map(item => item.value) : null,
     );
   }