+++ src/components/index.js
@@ -91,6 +91,7 @@ export EditTitleBlock from '@plone/volto/components/manage/Blocks/Title/Edit';
 export EditToCBlock from '@plone/volto/components/manage/Blocks/ToC/Edit';
 export EditTextBlock from '@plone/volto/components/manage/Blocks/Text/Edit';
 export EditImageBlock from '@plone/volto/components/manage/Blocks/Image/Edit';
+export EditListingBlock from '@plone/volto/components/manage/Blocks/Listing/Edit';
 export EditVideoBlock from '@plone/volto/components/manage/Blocks/Video/Edit';
 export EditBlock from '@plone/volto/components/manage/Blocks/Block/Edit';
 export EditHeroImageLeftBlock from '@plone/volto/components/manage/Blocks/HeroImageLeft/Edit';
@@ -103,6 +104,7 @@ export ViewTitleBlock from '@plone/volto/components/manage/Blocks/Title/View';
 export ViewToCBlock from '@plone/volto/components/manage/Blocks/ToC/View';
 export ViewTextBlock from '@plone/volto/components/manage/Blocks/Text/View';
 export ViewImageBlock from '@plone/volto/components/manage/Blocks/Image/View';
+export ViewListingBlock from '@plone/volto/components/manage/Blocks/Listing/View';
 export ViewVideoBlock from '@plone/volto/components/manage/Blocks/Video/View';
 export ViewMapBlock from '@plone/volto/components/manage/Blocks/Maps/View';
 export ViewHTMLBlock from '@plone/volto/components/manage/Blocks/HTML/View';
@@ -113,3 +115,5 @@ export PersonalTools from '@plone/volto/components/manage/Toolbar/PersonalTools'
 export More from '@plone/volto/components/manage/Toolbar/More';
 export Types from '@plone/volto/components/manage/Toolbar/Types';
 export Toast from '@plone/volto/components/manage/Toast/Toast';
+
+export ConditionalLink from '@plone/volto/components/manage/ConditionalLink/ConditionalLink';