+++ src/components/index.js
@@ -109,6 +109,10 @@ export ViewVideoBlock from '@plone/volto/components/manage/Blocks/Video/View';
 export ViewMapBlock from '@plone/volto/components/manage/Blocks/Maps/View';
 export ViewHTMLBlock from '@plone/volto/components/manage/Blocks/HTML/View';
 
+export ListingBlockBody from '@plone/volto/components/manage/Blocks/Listing/ListingBody';
+export ListingBlockData from '@plone/volto/components/manage/Blocks/Listing/ListingData';
+export ListingBlockSidebar from '@plone/volto/components/manage/Blocks/Listing/ListingSidebar';
+
 export ImageSidebar from '@plone/volto/components/manage/Blocks/Image/ImageSidebar';
 
 export PersonalTools from '@plone/volto/components/manage/Toolbar/PersonalTools';