+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changes
 
+- fix bug in TokenWidget and ArrayWidget when removing all elements. @giuliaghisini
+
 ## 4.0.0-alpha.29 (2020-01-18)
 
 ### Changes