+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### Added
 
+- Added listing block @pnicolli @rodfersou @sneridagh @giuliaghisini
+
 ### Changes
+
+- fix bug in TokenWidget and ArrayWidget when removing all elements. @giuliaghisini
+- fix listing block customization @giuliaghisini
 
 ## 4.0.0-alpha.29 (2020-01-18)