+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - fix bug in TokenWidget and ArrayWidget when removing all elements. @giuliaghisini
 - fix listing block customization @giuliaghisini
+- fix Querystring widgets failing to render if the `MultipleSelectionWidget` field applied is not included in the site vocabulary returned by `@querystring`. This applies to `Subject` field when the tag value is not there any more. @sneridagh
 
 ## 4.0.0-alpha.29 (2020-01-18)