+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 4.0.0 (unreleased)
+## 4.0.0-alpha.30 (2020-01-22)
 
 ### Added